working questions

+++ src/pages/pregunta/index.vue
@@ -8,10 +8,11 @@
       <a-alert
         v-else-if="currentSession"
         message="Sesión en curso"
-        :description="currentSession"
+        :description="eventTitle"
         type="info"
         style="font-size: 1rem"
         show-icon
+        class="max-w-[80%] mx-auto"
       />
       <a-alert
         v-else-if="error"
@@ -21,43 +22,89 @@
         show-icon
       />
       <a-alert
-        v-else="error"
+        v-else
         message="Actualmente no hay ninguna sesión en curso."
         type="warning"
         style="font-size: 1rem"
         show-icon
       />
     </div>
-    <main v-auto-animate class="flex flex-col flex-grow items-center md:justify-center">
-      <div v-if="currentSession" class="max-md:mt-10 w-6/12 max-w-[35rem] min-w-[20rem]">
-        <a-input-search
-          v-model:value="question"
-          placeholder="¿Conoces el chiste del pingüino?"
-          size="large"
-          class="w-full"
-          :status="error ? 'error' : ''"
-          :disabled="loading"
-          @search="onSubmit"
-        >
-          <template #enterButton>
-            <a-button type="primary">Preguntar</a-button>
-          </template>
-        </a-input-search>
-      </div>
+
+    <form
+      v-if="currentSession"
+      @submit.prevent="sendQuestion"
+      class="max-md:mt-10 w-6/12 max-w-[35rem] min-w-[20rem] mb-6 mx-auto flex flex-col items-stretch justify-center"
+    >
+      <a-textarea
+        v-model:value="questionToAsk"
+        placeholder="Escribe tu pregunta"
+        size="large"
+        class="w-full"
+        :status="error ? 'error' : ''"
+        :disabled="loading"
+        required
+        :minlength="20"
+        :maxlength="500"
+        showCount
+      >
+        <template #enterButton>
+          <a-button type="primary">Preguntar</a-button>
+        </template>
+      </a-textarea>
+      <a-button html-type="submit" type="primary" class="mt-5"> Enviar </a-button>
+    </form>
+
+    <main v-auto-animate class="">
+      <a-card
+        v-for="question in questions"
+        :key="question.id"
+        class="w-80 mb-4 mx-auto border-0"
+        :class="{
+          'border-l-8 border-primary': question.user_id === userStore.user?.id && !question.hidden,
+          'border-l-8 border-amber-500': question.hidden
+        }"
+        :body-style="{
+          display: 'flex',
+          flexDirection: 'row',
+          flexWrap: 'nowrap',
+          alignItems: 'stretch',
+          justifyContent: 'space-evenly',
+          padding: 0
+        }"
+        @contextmenu.prevent="() => promptDeleteQuestion(question.id, question.user_id ?? -1)"
+      >
+        <p class="whitespace-pre-wrap flex-grow max-w-[85%] break-words p-5">
+          {{ question.question }}
+        </p>
+        <div class="flex flex-col justify-end" @click="toggleLike(question.id)">
+          <a-badge :count="question.question_likes[0].count ?? 0" color="cyan"></a-badge>
+          <IconThumbUpFilled v-if="isQuestionLiked(question.id)" class="size-5 mb-2" />
+          <IconThumbUp class="size-5 mb-2" v-else />
+        </div>
+
+        <p class="absolute left-2 top-1 text-sm text-gray-500" v-if="question.hidden">
+          <IconEyeOff class="inline-block size-4 -mt-1.5" /> Oculta
+        </p>
+      </a-card>
     </main>
+
     <Footer />
   </div>
+
+  <contextHolder />
 </template>
 
 <script setup lang="ts">
-import { message } from 'ant-design-vue'
+import { message, Modal } from 'ant-design-vue'
 import { useUserStore } from '@/stores/user'
 import { useMECStore } from '@/stores/mec'
 import { useRouter } from 'vue-router'
 import { createClient } from '@supabase/supabase-js'
 import { useAsyncState } from '@vueuse/core'
+import type { Database } from '../../supabase-types'
+import { IconThumbUpFilled, IconThumbUp, IconAlertCircle, IconEyeOff } from '@tabler/icons-vue'
 
-const supabase = createClient(
+const supabase = createClient<Database>(
   import.meta.env.SUPABASE_URL_2024,
   import.meta.env.SUPABASE_KEY_2024,
   { auth: { persistSession: false } }
@@ -66,31 +113,34 @@ const router = useRouter()
 const userStore = useUserStore()
 const error = ref(false)
 const loading = ref(true)
-const question = ref('')
 const mec = useMECStore()
+
+const [modal, contextHolder] = Modal.useModal()
 
 const __fn = () =>
   supabase.from('question_sessions').select('*').eq('is_open', true).limit(1).maybeSingle()
-type Session = Awaited<ReturnType<typeof __fn>>
+type Session = Awaited<ReturnType<typeof __fn>>['data']
 
 mec.load()
 const { state: currentSession, execute: reloadSession } = useAsyncState(
   async () =>
-    await supabase.from('question_sessions').select('*').eq('is_open', true).limit(1).maybeSingle(),
+    await supabase
+      .from('question_sessions')
+      .select('*')
+      .eq('is_open', true)
+      .limit(1)
+      .maybeSingle()
+      .then(({ data }) => data),
   null as Session | null,
   {
-    resetOnExecute: false
+    resetOnExecute: false,
+    immediate: false,
+    onError(e) {
+      message.error('Error al cargar la sesión actual: ' + e)
+    }
   }
 )
-// onMounted hook is needed to ensure that setTimeout is defined
-onMounted(() => setInterval(reloadSession, 10000))
-const eventTitle = computed(
-  () => mec.mec?.findId(currentSession.value?.data?.event_id ?? -1)?.title
-)
-
-const onSubmit = () => {
-  console.log(question.value)
-}
+const eventTitle = computed(() => mec.mec?.findId(currentSession.value?.event_id ?? -1)?.title)
 
 watch(
   () => userStore.user,
@@ -101,22 +151,14 @@ watch(
   }
 )
 
-onMounted(() => {
-  if (!userStore.user) {
-    router.push('/pregunta/login')
-  }
-})
-
-const fn = async () =>
-  ((await supabase.from('questions').select('*, question_likes(count)')).data ?? [])
-    .sort((a, b) => b.question_likes[0].count - a.question_likes[0].count)
-    .filter((q) => q.user_id === userStore.user?.id || !q.hidden)
-type QS = Awaited<ReturnType<typeof fn>>
-
-const questions: Ref<QS> = ref([])
-;(async () => {
-  questions.value = (
-    (await supabase.from('questions').select('*, question_likes(count)')).data ?? []
+const fetchQuestions = async () =>
+  (
+    (
+      await supabase
+        .from('questions')
+        .select('*, question_likes(count)')
+        .eq('session_id', currentSession.value?.id ?? -1)
+    ).data ?? []
   )
     .sort(
       (a, b) =>
@@ -124,6 +166,11 @@ const questions: Ref<QS> = ref([])
         (a.question_likes ?? [{ count: 0 }])[0].count
     )
     .filter((q) => q.user_id === userStore.user?.id || !q.hidden)
+type QS = Awaited<ReturnType<typeof fetchQuestions>>
+
+const questions: Ref<QS> = ref([])
+const setupQuestions = async () => {
+  questions.value = await fetchQuestions()
   supabase
     .channel('supabase_realtime')
     .on(
@@ -134,6 +181,17 @@ const questions: Ref<QS> = ref([])
         table: 'questions'
       },
       (payload) => {
+        // Not working but shouldn't need this
+        // if (
+        //   payload.new &&
+        //   payload.new.session_id !== currentSession.value?.id &&
+        //   payload.old &&
+        //   payload.old.session_id !== currentSession.value?.id
+        // ) {
+        //   console.warn('Not for this session', payload)
+        //   return
+        // }
+
         if (payload.eventType === 'DELETE')
           questions.value.splice(
             questions.value.findIndex((q) => q.id === payload.old.id),
@@ -141,7 +199,7 @@ const questions: Ref<QS> = ref([])
           )
         else if (payload.eventType === 'INSERT') questions.value.push(payload.new as QS[number])
         else {
-          logger.info('updating question', payload)
+          console.info('updating question', payload)
           const idx = questions.value.findIndex((q) => q.id === payload.old.id)
 
           if (idx == -1) questions.value.push(payload.new as QS[number])
@@ -165,7 +223,12 @@ const questions: Ref<QS> = ref([])
       },
       async () => {
         questions.value = (
-          (await supabase.from('questions').select('*, question_likes(count)')).data ?? []
+          (
+            await supabase
+              .from('questions')
+              .select('*, question_likes(count)')
+              .eq('session_id', currentSession.value?.id ?? -1)
+          ).data ?? []
         )
           .sort(
             (a, b) =>
@@ -176,91 +239,68 @@ const questions: Ref<QS> = ref([])
       }
     )
     .subscribe()
-})()
+}
 
 const liked = ref([] as number[])
-;(async () => {
+const setupLiked = async () => {
   const my_likes = await supabase
     .from('question_likes')
     .select('*')
     .eq('user_id', userStore.user?.acf.id_base_de_datos_app ?? -1)
   liked.value = my_likes.data?.map((q) => q.question_id) ?? []
-})()
+}
+const isQuestionLiked = computed(() => (id: number) => liked.value.includes(id))
 
 const likesGiven: Record<number, number> = {}
 const areLikesGivenLoaded = ref(false)
 const computeLikesGiven = async () => {
   if (userStore.user === null) {
-    useToast({
-      color: 'danger',
-      message: t('question.modal.toasts.nullUserError')
-    })
-    dismissModal()
-  } else if (likesGiven[props.session.id] === undefined) {
+    message.error('No se ha encontrado el usuario')
+  } else if (currentSession.value === null) {
+    message.error('No se ha encontrado la sesión actual')
+  } else if (likesGiven[currentSession.value.id ?? -1] === undefined) {
     const { data, error } = await supabase
       .from('user_likes_given')
       .select('*', { count: 'exact' })
-      .eq('session_id', props.session.id)
+      .eq('session_id', currentSession.value.id)
       .eq('user_id', userStore.user.id)
     if (error) {
-      useToast({
-        color: 'danger',
-        message: t('question.modal.toasts.errorRetrievingUserLikes')
-      })
-      dismissModal()
+      message.error('Error al obtener los likes: ' + error.message)
     } else {
       if (data === null || data.length === 0 || data[0].count === 0) {
         const { error } = await supabase.from('user_likes_given').insert({
-          session_id: props.session.id,
+          session_id: currentSession.value.id,
           user_id: userStore.user.id,
           like_count: 0
         })
-        logger.info('likes inserted', error)
+        console.info('likes inserted', error)
       } else {
-        likesGiven[props.session.id] = data[0].like_count
+        likesGiven[currentSession.value.id] = data[0].like_count
       }
     }
     areLikesGivenLoaded.value = true
   }
 }
-onMounted(computeLikesGiven)
-watch(props.session, computeLikesGiven)
-
-// const dbg = <T,>(x: T) => {
-//   console.info('dbg', x);
-//   return x;
-// };
 
 const toggleLike = async (id: number, isInitial: boolean = false) => {
+  if (currentSession.value === null) {
+    message.error('No se ha encontrado la sesión actual')
+    return
+  }
+
   if (liked.value.includes(id)) {
     const { error } = await supabase
       .from('question_likes')
       .delete()
       .eq('user_id', userStore.user?.acf.id_base_de_datos_app ?? -1)
       .eq('question_id', id)
-    if (error)
-      useToast({
-        color: 'danger',
-        icon: warningOutline,
-        message: 'Error al registrar reacción a la pregunta',
-        cssClass: ''
-      })
+    if (error) message.error('Error al registrar reacción a la pregunta')
     else liked.value.splice(liked.value.indexOf(id), 1)
-  } else if (likesGiven[props.session.id] > 15 && !isInitial) {
-    useToast({
-      color: 'danger',
-      icon: closeCircleOutline,
-      message: t('question.modal.toasts.likeLimitReached'),
-      cssClass: ''
-    })
+  } else if (likesGiven[currentSession.value.id] > 15 && !isInitial) {
+    message.error('No puedes dar más a like en esta sesión')
   } else {
     if (userStore.user?.acf.id_base_de_datos_app === undefined) {
-      useToast({
-        color: 'danger',
-        icon: warningOutline,
-        message: t('question.modal.toasts.nullUserError'),
-        cssClass: ''
-      })
+      message.error('No se ha encontrado el usuario')
       return
     }
     const { error } = await supabase.from('question_likes').insert({
@@ -268,34 +308,18 @@ const toggleLike = async (id: number, isInitial: boolean = false) => {
       question_id: id
     })
     if (error) {
-      useToast({
-        color: 'danger',
-        icon: warningOutline,
-        message: t('question.modal.toasts.saveLikeError'),
-        cssClass: ''
-      })
+      message.error('Error al registrar la reacción')
     } else {
       liked.value.push(id)
       if (!isInitial) {
         const { error } = await supabase.rpc('increment_user_likes', {
           req_user_id: userStore.user.id,
-          req_session_id: props.session.id
+          req_session_id: currentSession.value.id
         })
         if (error) {
-          useToast({
-            color: 'warning',
-            cssClass: '',
-            icon: warningOutline,
-            message: t('question.modal.toasts.registerLikeError')
-          })
-          logPostgrestError(
-            'questionsModal:toggleLike',
-            'rpc(increment_user_likes) for ' + id.toString(),
-            null,
-            error
-          )
+          message.warn('Error al registrar la reacción')
         }
-        likesGiven[props.session.id] += 1
+        likesGiven[currentSession.value.id] += 1
       }
     }
   }
@@ -306,18 +330,17 @@ const publishingQuestion = ref(false)
 const sendQuestion = async () => {
   if (publishingQuestion.value) return
 
+  if (questionToAsk.value.length < 20) {
+    message.error('Mínimo 20 caracteres')
+    return
+  } else if (questionToAsk.value.length > 500) {
+    message.error('Máximo 500 caracteres')
+    return
+  }
+
   const questionParagraphs = questionToAsk.value.split('\n').length
   if (questionParagraphs > 3) {
-    useToast({
-      color: 'warning',
-      icon: warningOutline,
-      message: t('question.modal.toasts.maxLinesReached'),
-      cssClass: ''
-    })
-    logger.info(
-      'questionsModal:sendQuestion',
-      'user tried to ask more than 3 lines: ' + questionParagraphs
-    )
+    message.error('Máximo 3 párrafos por pregunta')
     return
   }
 
@@ -336,20 +359,8 @@ const sendQuestion = async () => {
   publishingQuestion.value = false
 
   if (error) {
-    if (error['custom_error_message'])
-      useToast({
-        color: 'danger',
-        icon: warningOutline,
-        message: error.custom_error_message,
-        cssClass: ''
-      })
-    else
-      useToast({
-        color: 'danger',
-        icon: warningOutline,
-        message: t('question.modal.toasts.genericSendError'),
-        cssClass: ''
-      })
+    if (error['custom_error_message']) message.error(error.custom_error_message)
+    else message.error('Error al enviar la pregunta')
     return
   }
 
@@ -357,13 +368,29 @@ const sendQuestion = async () => {
   toggleLike(data, true)
 }
 
+const promptDeleteQuestion = (id: number, question_user_id: number) => {
+  if (userStore.user?.id != question_user_id) return
+
+  console.info('hi')
+  modal.confirm({
+    title: '¿Seguro que quieres borrar esta pregunta?',
+    icon: h(IconAlertCircle, { class: 'size-10 inline-block text-amber-500' }),
+    content: 'Esta acción no puede ser revertida. Los likes de la pregunta se perderán.',
+    okText: 'Eliminar',
+    okType: 'danger',
+    cancelText: 'Cancelar',
+    onOk() {
+      deleteQuestion(id, question_user_id)
+    },
+    onCancel() {
+      console.log('Cancel')
+    }
+  })
+}
 const deleteQuestion = async (id: number, question_user_id: number) => {
   if (userStore.user?.id != question_user_id) return
 
-  const progressToast = await useToast({
-    message: t('question.modal.toasts.deletingQuestionInProgress'),
-    cssClass: ''
-  })
+  const hideLoading = message.loading('Borrando pregunta...', 0)
 
   const { error } = await supabase
     .from('questions')
@@ -371,31 +398,37 @@ const deleteQuestion = async (id: number, question_user_id: number) => {
     .eq('id', id)
     .eq('user_id', userStore.user.id)
 
-  progressToast.dismiss()
+  hideLoading()
   if (error) {
-    useToast({
-      color: 'danger',
-      message: t('question.modal.toasts.deletingQuestionError'),
-      icon: closeCircleOutline,
-      cssClass: ''
-    })
-    logPostgrestError(
-      'questionsModal:deleteQuestion',
-      'error from supabase when deleting question' + error.message,
-      null,
-      error
-    )
+    message.error('Error al eliminar la pregunta')
   } else {
-    useToast({
-      color: 'success',
-      message: t('question.modal.toasts.deletingQuestionSucceded'),
-      icon: checkmarkCircleOutline,
-      cssClass: ''
-    })
+    message.success('Pregunta eliminada correctamente')
   }
 }
 
+const load = async (loadSession = true) => {
+  loading.value = true
+
+  if (loadSession) await reloadSession()
+  if (currentSession.value !== null) {
+    await setupQuestions()
+    await setupLiked()
+    await computeLikesGiven()
+  }
+
+  loading.value = false
+}
+
+watch(currentSession, (newSession, oldSession) => {
+  if (newSession && newSession.id !== oldSession?.id) load(false)
+})
+
 onMounted(() => {
-  logger.trace('questionsModal:onMounted', 'current session', props.session)
+  if (!userStore.user) {
+    router.push('/pregunta/login')
+  }
+
+  load()
+  setInterval(reloadSession, 10000)
 })
 </script>
